plugins/networks: implemented Connman authentication agent
The new Connman API no longer makes it possible to pass connection parameters
of any sort with the 'Connect' request. Instead Connman expects there to be
an authentication agent which will provide whatever is need on request.

The agent implementation also adds support for networks that require a
username (e.g., EAP-PEAP, EAP-TTLS), though how well it works was not tested
(there is only so much Connman a normal person can take in space of 24h).